@@ -10,25 +10,25 @@ class ApiErrorHandler {
 
     if (error is Exception) {
       try {
-        if (error is DioError) {
+        if (error is DioException) {
 
           switch (error.type)
           {
-            case DioErrorType.cancel:
+            case DioExceptionType.cancel:
               errorDescription = "Request to API server was cancelled";
               break;
-            case DioErrorType.connectTimeout:
+            case DioExceptionType.connectionTimeout:
               errorDescription = "Connection timeout with API server";
               break;
-            case DioErrorType.other:
+            case DioExceptionType.unknown:
               errorDescription =
               "Connection to API server failed due to internet connection";
               break;
-            case DioErrorType.receiveTimeout:
+            case DioExceptionType.receiveTimeout:
               errorDescription =
               "Receive timeout in connection with API server";
               break;
-            case DioErrorType.response:
+            case DioExceptionType.badResponse:
               switch (error.response!.statusCode)
               {
                 case 404:
@@ -49,8 +49,14 @@ class ApiErrorHandler {
                   }
               }
               break;
-            case DioErrorType.sendTimeout:
+            case DioExceptionType.sendTimeout:
               errorDescription = "Send timeout with server";
+              break;
+            case DioExceptionType.badCertificate:
+               errorDescription = "badCertificate";
+              break;
+            case DioExceptionType.connectionError:
+               errorDescription = "connectionError";
               break;
           }
         }
