@@ -22,8 +22,8 @@ class DioClient {
     dio = dioC;
     dio
     !..options.baseUrl = baseUrl
-      ..options.connectTimeout = 90000
-      ..options.receiveTimeout = 90000
+      ..options.connectTimeout =  Duration(seconds: 90)
+      ..options.receiveTimeout =  Duration(seconds: 90)
       ..httpClientAdapter
       ..options.headers = {
         'Content-Type': 'application/json; charset=UTF-8',
